scraping_tool/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
async-timeout==5.0.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
fastapi==0.115.6
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
parse==1.20.2
pydantic==2.10.3
pydantic_core==2.27.1
pyee==11.1.1
Pygments==2.18.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-html==0.10.0
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.2.1
watchfiles==1.0.3
websockets==10.4
zipp==3.21.0
